pyproject.toml
[tool.poetry]
name = "stable-diffusion-server"
version = "0.1.0"
description = ""
authors = ["Rafael Irgolic <hello@irgolic.com>"]
license = "GNU Affero General Public License"
readme = "README.md"
packages = [{include = "stable_diffusion_server"}]

[tool.poetry.dependencies]
python = "^3.10"
python-jose = "^3.3.0"
PyYAML = "^6.0"
requests = "^2.28.1"
requests-toolbelt = "^0.10.0"
fastapi = "^0.85.0"
typing-extensions = "^4.4.0"
passlib = "^1.7.4"
redis = "^4.3.4"
aioredis = "^2.0.1"
python-multipart = "^0.0.5"
uvicorn = "^0.18.3"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.274"


[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"
pytest-asyncio = "^0.19.0"
websockets = "^10.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
